Moved repeater custom fields from brick builder back to parent field builder class so the collction builder has support
